requirements.txt
# cat requirements.txt

appdirs==1.4.4
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
idna==2.10
lxml==4.5.2
msgpack==1.0.0
numpy==1.19.2
pipenv==2020.8.13
py-algorand-sdk==1.4.0
pycparser==2.20
pycryptodomex==3.9.8
PyJWT==1.7.1
PyNaCl==1.4.0
pyteal==0.6.0
python-dotenv==0.15.0
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
twilio==6.45.4
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.0.35
virtualenv-clone==0.5.4
yapf==0.30.0